Refactor configureStore and base redux types and added a single reselect selector
